Kaizen: Move OffTopicName to converters.py.
