+++ bot/cogs/off_topic_names.py
@@ -4,44 +4,17 @@ import logging
 from datetime import datetime, timedelta
 
 from discord import Colour, Embed
-from discord.ext.commands import BadArgument, Cog, Context, Converter, group
+from discord.ext.commands import Cog, Context, group
 
 from bot.api import ResponseCodeError
 from bot.bot import Bot
 from bot.constants import Channels, MODERATION_ROLES
+from bot.converters import OffTopicName
 from bot.decorators import with_role
 from bot.pagination import LinePaginator
 
-
 CHANNELS = (Channels.off_topic_0, Channels.off_topic_1, Channels.off_topic_2)
 log = logging.getLogger(__name__)
-
-
-class OffTopicName(Converter):
-    """A converter that ensures an added off-topic name is valid."""
-
-    @staticmethod
-    async def convert(ctx: Context, argument: str) -> str:
-        """Attempt to replace any invalid characters with their approximate Unicode equivalent."""
-        allowed_characters = "ABCDEFGHIJKLMNOPQRSTUVWXYZ!?'`-"
-
-        # Chain multiple words to a single one
-        argument = "-".join(argument.split())
-
-        if not (2 <= len(argument) <= 96):
-            raise BadArgument("Channel name must be between 2 and 96 chars long")
-
-        elif not all(c.isalnum() or c in allowed_characters for c in argument):
-            raise BadArgument(
-                "Channel name must only consist of "
-                "alphanumeric characters, minus signs or apostrophes."
-            )
-
-        # Replace invalid characters with unicode alternatives.
-        table = str.maketrans(
-            allowed_characters, '𝖠𝖡𝖢𝖣𝖤𝖥𝖦𝖧𝖨𝖩𝖪𝖫𝖬𝖭𝖮𝖯𝖰𝖱𝖲𝖳𝖴𝖵𝖶𝖷𝖸𝖹ǃ？’’-'
-        )
-        return argument.translate(table)
 
 
 async def update_names(bot: Bot) -> None:
