+++ bot/converters.py
@@ -237,6 +237,32 @@ class Duration(DurationDelta):
             raise BadArgument(f"`{duration}` results in a datetime outside the supported range.")
 
 
+class OffTopicName(Converter):
+    """A converter that ensures an added off-topic name is valid."""
+
+    async def convert(self, ctx: Context, argument: str) -> str:
+        """Attempt to replace any invalid characters with their approximate Unicode equivalent."""
+        allowed_characters = "ABCDEFGHIJKLMNOPQRSTUVWXYZ!?'`-"
+
+        # Chain multiple words to a single one
+        argument = "-".join(argument.split())
+
+        if not (2 <= len(argument) <= 96):
+            raise BadArgument("Channel name must be between 2 and 96 chars long")
+
+        elif not all(c.isalnum() or c in allowed_characters for c in argument):
+            raise BadArgument(
+                "Channel name must only consist of "
+                "alphanumeric characters, minus signs or apostrophes."
+            )
+
+        # Replace invalid characters with unicode alternatives.
+        table = str.maketrans(
+            allowed_characters, '𝖠𝖡𝖢𝖣𝖤𝖥𝖦𝖧𝖨𝖩𝖪𝖫𝖬𝖭𝖮𝖯𝖰𝖱𝖲𝖳𝖴𝖵𝖶𝖷𝖸𝖹ǃ？’’-'
+        )
+        return argument.translate(table)
+
+
 class ISODateTime(Converter):
     """Converts an ISO-8601 datetime string into a datetime.datetime."""
